diff: keep track of the type of the last line seen
The "\ No newline at the end of the file" can come after any of the
"-" (deleted preimage line), " " (unchanged line), or "+" (added
postimage line).  In later steps in this series, we will start
treating a change that makes a file to end in an incomplete line,
and we would need to know what the previous line was when we react
to "\ No newline" in the diff output.  If the previous line was a
context (i.e., unchanged) line, the file lacked the final newline
before the change, and the change did not touch that line and left
it still incomplete, so we do not want to warn in such a case, for
example.

Teach fn_out_consume() function to keep track of what the previous
line was, and prepare an otherwise empty switch statement to let us
react differently to "\ No newline" based on that.

Note that there is an existing curiosity (read: likely to be a bug)
in the code that increments line number in the preimage file every
time it sees a line with "\ No newline" on it, regardless of what
the previous line was.  I left it as-is, because it does not affect
the main theme of this series, and more importantly I do not think
it matters, as these numbers are used only to compare them with
blank_at_eof_in_{pre,post}image to issue the warning every time we
see an added line, but by definition, after we see "\ No newline at
the end of the file" for an added line, we will not see an added
line for the file.  As a #leftoverbit task, an independent audit to
ensure that this curious increment can be safely removed would make
a good clean-up (we may even find some code that decrements this
counter or over-increments the other quantity this counter is
compared with that compensates the effect of this curious increment
that hides a bug, in which case we may also need to remove them).

Signed-off-by: Junio C Hamano <gitster@pobox.com>